Basic setup and class initialization
Setup the index.html with boilerplate.
Create a main.js and add it in the
script tag in index.html so that
main.js can be loaded as a module.

Create a linked-lists.js and import it
to main.js.

In linked-lists.js, create and export
the LinkedList and Node classes.

Create a new instance of the
LinkedList class in main.js