Implement append()
In the LinkedList class, implement
append(), a method which takes in
a `value` and adds a new node
containing `value` to the
end of the list.

The method traverses the list from
the head and adds the new node to
the end of the list.

In main.js, append four lists
and log the list to the console.