Implement size()
In the LinkedList class, implement
size() which returns the total
number of nodes in the list.

The method sets a counter variable
and traverses from the head to
the last node, updating the counter
variable in the process. Return
the counter variable.

In main.js, log the results of
calling the size() method.