Implement getNodeAt().
Take in an index argument.
Check if index is less than 1.
check if greater than total
number of nodes.

If index is from 1 to the size
of the list, traverse through
the list from
the head until
the node at the given index is
found. Then return the node.

In main.js, log the results of
getNodeAt(3).