Implement pop() in LinkedList
Remove the last element
in the list and return it.

If nodeSize is 0, return

Get the lastNode and check
if nodeSize is 1, if it is,
set the head to null and
return the lastNode.

If there is more than one
node, get the secondToLastNode,
set its nextNode to null and
return lastNode.

In main.js, call the pop() and check the list size.