Implement insertAt()
The method takes in a `value`
and an `index` and inserts
a new node with `value` at
`index`. Also return the
new inserted node.

Make the new node the head,
if the list was empty.

If the `index` is 1, make
the new node the head.

if the index is larger than
the list size, make the new
node the last node(tail)

Else, insert the node
at the index.

In main.js, log the result
of calling the method
with args.