Update insertAt() and implement removeAt()
Update insertAt() to return
if the specified `index` is
less than 1.

Implement removeAt() as a method
that takes in an `index`,
removes the node at the
specified index and returns
the removed node.

Return if the index is less than
1 or if it is greater than the
listSize.

If the index is 1 and the listSize
is 1, remove the head node
and return it

If the index is 1, make the
second node, the new head node.

Else, get the previous node
of the node at specified index,
set its nextNode to be the
next node of the node at
the specified index.
Set the nextNode of the
node at the specified index
to be null. C'est tout.

In main.js, log the result
of calling the removeAt()
method with an arg.